Updated the sssp code to support int32/uint64,float.  See the nodes in sssp.h for how to compile for these different environments. There is also an issue with atomicMin() for int64 requiring cuda compute 3.5+.

+++ Algorithms/SSSP/sssp.cu
@@ -445,7 +445,7 @@ int main(int argc, char **argv)
     FILE* f = fopen(outFileName, "w");
     for (int i = 0; i < csr_graph.nodes; ++i)
     {
-      fprintf(f, "%d\n", h_values[i]);
+      fprintf(f, "%lld\n", (long long) h_values[i]);
     }
 
     fclose(f);
